Break up Format.kt
